socket connection refactor
